+++ SwiftXPC/Objects/Data.swift
@@ -11,7 +11,7 @@ import XPC
 
 public class XPCData : XPCObject {
     public convenience init(blob: NSData) {
-        self.init(nativePointer: xpc_data_create(blob.bytes, UInt(blob.length)))
+        self.init(nativePointer: xpc_data_create(blob.bytes, blob.length))
     }
     
     public convenience init(text: String) {
@@ -25,7 +25,7 @@ public class XPCData : XPCObject {
     
     public func readData(offset: Int = 0) -> NSData? {
         if let data = NSMutableData(length: self.length) {
-            let aCopied = xpc_data_get_bytes(objectPointer, data.mutableBytes, UInt(offset), UInt(self.length))
+            let aCopied = xpc_data_get_bytes(objectPointer, data.mutableBytes, offset, self.length)
             data.length = Int(aCopied)
             return NSData(data: data)
         } else {
